modify skimage data fixtures
